Add other_version_patterns config
This is to control how versions are replaced in the
other_version_file_paths. It is not good enough to replace version-A
with version-B in those files because doing so is likely to replace
an irrelevant version that just happens to match the same string.

To keep this working properly, we need some way to target relevant
versions only in those extra version files. In the spec added here the
extra version file is a Gemfile.lock, which references the main library
version in several ways (it's the same structure as ollyphaunt which
ships 2 gems, so it has 2 libraries with the same version - ollyphaunt
and ollyphaunt-railtie, both of those need to be changed in the lock
file to match the main version or the PR can't be merged).

The patterns are literal strings, not regular expressions, to keep things
as simple as possible. That way we don't need massive regexes in each
one for matching the version - we just use a convention that '1.2.3'
denotes where the version should be in the pattern, it will be replaced
by the actual versions involved

I've chosen / as a separator between these patterns because it's unlikely
to be contained within a pattern, whereas a comma might be (I can think
of a case of that - but it's pretty marginal).